modularized the models
