+++ db/index.js
@@ -3,7 +3,7 @@ const mongoose = require('mongoose');
 mongoose.Promise = global.Promise;
 
 //database connection
-mongoose.connect('mongodb://localhost/FlateShare');
+mongoose.connect('mongodb://localhost/flateShare');
 
 
 module.exports = mongoose;
