+++ models/index.js
@@ -1,0 +1,9 @@
+const mongoose = require('../db');
+const Todo = require('./itemSchema');
+const initialTotal = require('./totalSchema');
+
+
+module.exports = {
+  Todo,
+  initialTotal
+}
